requirements.txt
Flask
nltk
gunicorn
flask
firebase_admin
flask_cors
requests
python-dotenv
textblob
werkzeug